+++ tests/test_baseapi.py
@@ -46,6 +46,7 @@ class TestBaseAPIInitOptions(object):
         assert baseapi.url_path == ''
         assert baseapi.username is None
         assert baseapi.password is None
+        assert baseapi.metric_api_version is 'v2'
 
     def test_host(self):
         api = pypuppetdb.api.BaseAPI(host='127.0.0.1')
@@ -136,6 +137,18 @@ class TestBaseAPIInitOptions(object):
                                      password='password123')
         assert api.username == 'puppetdb'
         assert api.password == 'password123'
+
+    def test_metric_api_version_v1(self):
+        api = pypuppetdb.api.BaseAPI(metric_api_version='v1')
+        assert api.metric_api_version == 'v1'
+
+    def test_metric_api_version_v2(self):
+        api = pypuppetdb.api.BaseAPI(metric_api_version='v2')
+        assert api.metric_api_version == 'v2'
+
+    def test_metric_api_version_invalid_raises(self):
+        with pytest.raises(ValueError):
+            pypuppetdb.api.BaseAPI(metric_api_version='bad')
 
 
 class TestBaseAPIProperties(object):
@@ -344,6 +357,31 @@ class TestAPIQuery(object):
         httpretty.disable()
         httpretty.reset()
 
+    def test_with_payload_get(self, baseapi):
+        httpretty.enable()
+        stub_request('http://localhost:8080/pdb/query/v4/nodes')
+        baseapi._query('nodes',
+                       payload={'foo': 'bar'},
+                       count_by=1)
+        assert httpretty.last_request().querystring == {
+            'foo': ['bar'],
+            'count_by': ['1']}
+        httpretty.disable()
+        httpretty.reset()
+
+    def test_with_payload_post(self, baseapi):
+        httpretty.enable()
+        stub_request('http://localhost:8080/pdb/query/v4/nodes',
+                     method=httpretty.POST)
+        baseapi._query('nodes',
+                       payload={'foo': 'bar'},
+                       count_by=1,
+                       request_method='POST')
+        assert httpretty.last_request().body == json.dumps({'foo': 'bar',
+                                                            'count_by': 1}).encode("latin-1")
+        httpretty.disable()
+        httpretty.reset()
+
     def test_response_empty(self, baseapi):
         httpretty.enable()
         httpretty.register_uri(httpretty.GET,
@@ -430,7 +468,29 @@ class TestAPIQuery(object):
 
 
 class TestAPIMethods(object):
-    def test_metric(self, baseapi):
+    def test_metric_v1(self, baseapi):
+        httpretty.enable()
+        httpretty.enable()
+        stub_request('http://localhost:8080/metrics/v1/mbeans/test')
+        baseapi.metric('test', version='v1')
+        assert httpretty.last_request().path == '/metrics/v1/mbeans/test'
+
+    def test_metric_v1_list(self, baseapi):
+        httpretty.enable()
+        httpretty.enable()
+        stub_request('http://localhost:8080/metrics/v1/mbeans')
+        baseapi.metric(version='v1')
+        assert httpretty.last_request().path == '/metrics/v1/mbeans'
+
+    def test_metric_v1_version_constructor(self):
+        api = pypuppetdb.api.BaseAPI(metric_api_version='v1')
+        httpretty.enable()
+        httpretty.enable()
+        stub_request('http://localhost:8080/metrics/v1/mbeans/test')
+        api.metric('test')
+        assert httpretty.last_request().path == '/metrics/v1/mbeans/test'
+
+    def test_metric_v2(self, baseapi):
         metrics_body = {
             'request': {
                 'mbean': 'test:name=Num',
@@ -452,7 +512,52 @@ class TestAPIMethods(object):
         httpretty.disable()
         httpretty.reset()
 
-    def test_metric_error(self, baseapi):
+    def test_metric_v2_version_constructor(self):
+        api = pypuppetdb.api.BaseAPI(metric_api_version='v2')
+        metrics_body = {
+            'request': {
+                'mbean': 'test:name=Num',
+                'type': 'read'
+            },
+            'value': {
+                'Value': 0
+            },
+            'timestamp': 0,
+            'status': 200
+        }
+
+        httpretty.enable()
+        httpretty.register_uri(httpretty.GET, 'http://localhost:8080/metrics/v2/read/test:name=Num',
+                               body=json.dumps(metrics_body))
+        metric = api.metric('test:name=Num')
+        assert httpretty.last_request().path == '/metrics/v2/read/test%3Aname%3DNum'
+        assert metric['Value'] == 0
+        httpretty.disable()
+        httpretty.reset()
+
+    def test_metric_v2_version_string(self, baseapi):
+        metrics_body = {
+            'request': {
+                'mbean': 'test:name=Num',
+                'type': 'read'
+            },
+            'value': {
+                'Value': 0
+            },
+            'timestamp': 0,
+            'status': 200
+        }
+
+        httpretty.enable()
+        httpretty.register_uri(httpretty.GET, 'http://localhost:8080/metrics/v2/read/test:name=Num',
+                               body=json.dumps(metrics_body))
+        metric = baseapi.metric('test:name=Num', version='v2')
+        assert httpretty.last_request().path == '/metrics/v2/read/test%3Aname%3DNum'
+        assert metric['Value'] == 0
+        httpretty.disable()
+        httpretty.reset()
+
+    def test_metric_v2_error(self, baseapi):
         metrics_body = {
             'request': {
                 'mbean': 'test:name=Num',
@@ -471,6 +576,62 @@ class TestAPIMethods(object):
         assert httpretty.last_request().path == '/metrics/v2/read/test%3Aname%3DNum'
         httpretty.disable()
         httpretty.reset()
+
+    def test_metric_v2_escape_special_characters(self, baseapi):
+        metrics_body = {
+            'request': {
+                'mbean': 'test:name=Num',
+                'type': 'read'
+            },
+            'value': {
+                'Value': 0
+            },
+            'timestamp': 0,
+            'status': 200
+        }
+
+        httpretty.enable()
+        metric_name = 'test:special/chars!metric"name'
+        metric_escaped = 'test:special!/chars!!metric!"name'
+        metric_escaped_urlencoded = 'test%3Aspecial%21/chars%21%21metric%21%22name'
+        httpretty.register_uri(httpretty.GET,
+                               ('http://localhost:8080/metrics/v2/read/' + metric_escaped),
+                               body=json.dumps(metrics_body))
+        metric = baseapi.metric(metric_name)
+        assert httpretty.last_request().path == ('/metrics/v2/read/' + metric_escaped_urlencoded)
+        assert metric['Value'] == 0
+        httpretty.disable()
+        httpretty.reset()
+
+    def test_metric_v2_list(self, baseapi):
+        # test metric() (no arguments)
+        metrics_body = {
+            'request': {
+                'type': 'list'
+            },
+            'value': {
+                'java.util.logging': {
+                    'type=Logging': {
+                    }
+                },
+            },
+            'timestamp': 0,
+            'status': 200
+        }
+
+        httpretty.enable()
+        httpretty.register_uri(httpretty.GET,
+                               'http://localhost:8080/metrics/v2/list',
+                               body=json.dumps(metrics_body))
+        metric = baseapi.metric()
+        assert httpretty.last_request().path == '/metrics/v2/list'
+        assert metric == {'java.util.logging': {'type=Logging': {}}}
+        httpretty.disable()
+        httpretty.reset()
+
+    def test_metric_bad_version(self, baseapi):
+        with pytest.raises(ValueError):
+            baseapi.metric('test', version='bad')
 
     def test_facts(self, baseapi):
         facts_body = [{